oper_swap() now changes pointers, instead of data.
OLD: swap the data between the two nodes, without actually moving the
nodes themselves in the list.
NEW: swap what the top two nodes (and the two nodes surrounding them)
are pointing to, without touching any other members inside of each node.

PROBLEMS WITH OLD IMPLEMENTATION: The same problem I tried to address
with the change to the push operation; keeping the data and it's
associated address the same might make it easier to keep track of them
in the future.
MORE IMPORTANTLY: The struct for my linked list currently only contains
one data member besides the pointers to the prev and next nodes. So the old
implementation worked. However, when adding more members or modifying the
struct in any other way, the old implementation would force me to modify
oper_swap() accordingly. Only changing the pointers instead of the data
within the nodes means I'm free to modify the struct without changing
this fucntion. Yay for portability!