Can count elements in LIS, but not retrieve them
The creation of piles and the counting of said piles is going smoothly,
I always get the output I want. Retrieving the elements that the LIS
consists of with pointers isn't quite going the way I'd expect it to.

Currently it points to other elements seemingly at random.
Something for tomorrow to look at!