Fixed a bug, see notes.
So there's the initial LIS searching algorithm, where we only searched for the
LIS once from the starting index. In that case it can be assumed that
the first element in the LIS we encounter will also be the lowest
element in it.
Now that the LIS can start from any number in the sequence it's not
guaranteed that the first element marked as being part of the LIS is
also the lowest. So I had to add an additional check to ensure the
minimum value gets updated if the min value is further along the list.